requirements.txt
django
mysqlclient
dotenv
djangorestframework